backend/requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
python-dotenv==1.0.0
openai==0.27.4